update headers for variable granularity
